+++ HR-AUTOMATION.Application/InputModels/ProfileInputModel.cs
@@ -30,12 +30,12 @@ public class ProfileInputModel
     /// <summary>
     /// Gets or sets the min experience.
     /// </summary>
-    public int? MinExperience { get; set; }
+    public decimal? MinExperience { get; set; }
 
     /// <summary>
     /// Gets or sets the max experience.
     /// </summary>
-    public int? MaxExperience { get; set; }
+    public decimal? MaxExperience { get; set; }
 
     /// <summary>
     /// Gets or sets the profile name.
